change onboarding route
