+++ src/app/sessions/book/confirmation/page.tsx
@@ -2,9 +2,9 @@
  * @author Angela Guo <aguo921@aucklanduni.ac.nz>
  */
 
+import { ConfirmationBanner } from "@/components/icons";
 import { ConfirmationIcon } from "@/components/payment/ConfirmationIcon";
 import { ConfirmedSessionCard } from "@/components/payment/ConfirmedSessionCard";
-import { ConfirmationBanner } from "@/lib/icons";
 
 const sessions = [
   {
